requirements.txt
# Core ML & Data
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.0

# API Framework
fastapi==0.103.2
uvicorn[standard]==0.23.2
pydantic==2.6.4

# MLflow
mlflow==2.9.2

# Monitoring & Drift
evidently==0.4.15
scipy==1.11.4
plotly==5.18.0

# Airflow
apache-airflow==2.7.3
apache-airflow-providers-docker==3.8.0
psycopg2-binary==2.9.9
sqlalchemy==1.4.51

# Utilities
python-dotenv==1.0.0
joblib==1.3.2
requests==2.31.0
typing-extensions==4.8.0